Support coverage instrumentation on optimized code.
This isn't a super important usecase, but I was curious why instrumentation was broken and it turned out it was a classic case of unintended property renaming.

coverage instrumentation works by populating a special object called '__jscov' on the window object with coverage metadata.  This means that we need to ensure that both __jscov itself and its various sub properties are not renamed since the code that ultimately reads this data is not part of the same compilation unit.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=200777314