Refactor some typechecking logic for getprops
This is so we can reuse it in a future CL when checking property access in for-of loop initializers.

The only non-refactoring change is that we back off some checks when assigning to ".prototype" that we didn't before, which resolves an old TODO in TypeCheckTest.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=201705228