Add support for code splitting and many other flags to the GWT version.
All of AbstractCommandLineRunner was marked as GWTIncompatible. Adds GwtIncompatible annotations to each instance method and field and refactors others to be static so they are usable in the GWT runner. In addition, flag logic in the GWT runner was audited and made to much more closely match the logic of the CommandLineRunner.

Closes https://github.com/google/closure-compiler/pull/2979

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=202029643