Remove disableScriptFeatureValidation() in all checks and transpilation pass unit tests
The only places still disabling validation are optimization tests. I'm not sure it will be feasible to support this well in optimization passes, because optimizations can take an arbitrary expression or statement with arbitrary 'Features' and move it to a different file. (e.g. in CrossModuleCodeMotion and InlineFunctions)

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=202975908