Allow accessing undeclared variables using `typeof SomeVar`.
`typeof x === 'undefined'` is the strict mode compliant way to check if a
variable.

Users should ideally declare externs for variables, but this isn't possible for
TypeScript. TypeScript emits code snippets under certain situations that check
for the presence of a variable and use it if it is defined. We could suppress
`checkVars` in the emit, but we'd rather not, given the check is valuable.

Accepting use within an explicit `typeof` check seems unlikely to permit genuine
errors, and likely to just match explicit checks by users.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=203094773