Expands the tests of `InferJSDocInfo` to better clarify what the class does and prepare for ES6 support.
There's a bit of an issue in that the old tests were very unclear, but this update attempts to retain coverage as best as possible.

This change also fixes what appears to be a bug in handling enum types. Previously, `instanceof` was being used to detect them, which isn't correct when considering proxy types.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=203146643