Enhance scope class in SymbolTable to contain its index in parent scope.
This is needed to generate unique names for variables. Today we achieve that by treating each block scope as an anonymous function scope and generating "function%n" symbol for all block scopes. That leads to a lot of unnecessary symbols which are "not clickable" and generally misleading.

Instead "function%n" will be generated only for actual function scopes. But we still need to generate unique names for variables and block indices will help with that: we prepend containing scope indices to variable name making it unique.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=204334299