Allow duplicate @suppress tags by collecting the union of suppressions.
This makes it much easier for automated tools to add to the list of suppressions for a file without having to parse and merge JSDoc, just by appending to an existing @fileoverview comment.

This reduces errors reported from human authored code, but it seems unlikely that a duplicate @suppress tag is an actual mistake – human authored code usually has all suppress tags in the same line in a short comment, so it seems unlikely that anybody would accidentally add a dupe. On top of that, it seems fairly clear that taking the union of suppressions is the intended behaviour when having multiple @suppress tags.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=204550634