Type check uses of `super` correctly in static methods
This initial support uses a partial, hacky workaround in order to exactly reproduce the same warnings as type checking pre-transpiled code.  A more correct implementation would be based on the type of `this` in the static method, but currently `this` in static methods are treated as unknown, so tightening that would introduce many new errors across the codebase, which would block moving class transpilation to after type checking.  Instead, we will tighten the types in a later step.

This is the last known blocker for type checking classes natively.  Once this is in, we should be able to move class transpilation to after type checking.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=204554183