Fix bug with extending forward-referenced classes
This can happen if an IIFE defines the superclass on a global qualified name.  Previously it gave up when the name was not found in the scope, and so made a class with Object as its superclass.  We now make a NamedType for a missing qualified name superclass, which is more consistent with how it was handled in the past.  It's still not a good idea to do this, since we're unable to infer the overrides correctly, but at least it understands the prototype chain now.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=205000839