Support multiline ES6 import statements
Prior to this change, due to the line parsing nature of
JsFileLineParser, the parser would fail to match import/export
statements that span multiple lines.

This change tracks prior lines when JsFileParser encounters an import or
export statement without a terminating semicolon, then parses the set of
lines when a semicolon is encountered.

While JsFileParser is tracking prior lines, it will prevent the parser
from becoming shortcutted.

resolves #3025

This implementation uses the following features:

- Exit multi-line read of ES6 imports early

In order to avoid consuming the entire file in the case that the file
contains no semicolons; JsFileParser will stop reading lines when it
encounters a line that begins (ignoring whitespace) with a commonly used
keyword.

The keywords were choosen based on a relatively small (self) selected
sample of JS code found on npmjs.org, using the following command:

    find src -name '*.js' -type f | head -100000 | while read f; do
      sed -n '/^\(ex\|im\)port/,$p' $f \
      | awk '{print$1}' \
      | grep -vE '/^\s*$/';
    done \
    | head -1000000 | sort \
    | uniq -c | sort -n

- Limit multi-line search by character count

This change imposes a limit of the number of characters read by
JsFileParser. When this limit is exceeded the search exits early and
raises a compiler warning.

The limit value was choosen as the power of two less the length of a
reasonable line of code (100) that is greater than or equal to the
longest import statement found in a relatively small self-selected
sample of ES6 code found on npmjs.org and other sources. This code was
found in the index.js file of the d3-time project (from the utcWeek
module).