Fixes #2605 - CommonJS missing semicolon check.
`ProcessCommonJSModules` creates a top-level 'initModule' var that
represents the exported module. The initModule `Node` is given
source info from the entire script node.

When the script being processed contains a newline after the last
statement, the erroneous missing semicolon is emitted since it is
not expecting the source information of a `var` to include a new
line.

The most minimal example:
```
module.exports = 1;
```
Compiled with `--jscomp_error=lintChecks --process_common_js_modules`