Fix crash when writing `class Foo extends Foo {}`
This is obviously an error, but the compiler shouldn't crash on it.  In fixing this, I also cleaned up some variable names in `TypeValidator#expectExtends` since it was difficult to tell what the difference between "superCtor" and "declaredSuper" were.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=206108306