Fix a ClassCastException in InferJSDocInfo
When a class node has an UnknownType rather than a FunctionType (this can happen when the class is immediately inside a CAST node to unknown, though I'm not positive that that's the root cause), trying to cast the type to FunctionType fails.  In general, casting JSType is almost always the wrong thing to do.  Instead, JSType.toMaybeFunctionType is much safer.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=206239547