Fix inference of super() calls
Previously we were inferring them based on the return type of the constructor, which is typically undefined.  But it actually returns the 'this' instance for the remainder of the constructor, and we should infer it using the new: slot of the function.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=206413799