Only do const inference on string keys in destructuring in TypedScopeCreator
This is closer to what we do right now for non-destructuring const declarations - we type literals and qualfied names, but not GETELEMs.

Also, it really doesn't make sense to try to infer types for objects with arbitrary default values in TypedScopeCreator, because we may not know what the default value type is until TypeInference runs.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=207574238