Update `CheckAccessControls` to bring checks on class-syntax up to parity with old-style classes.
This change isn't intended to improve coverage of the checks, which are known to miss some cases. Such a change would be much harder to land and require more thought as to what the correct behaviour is in some ambiguous cases.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=208274192