Work around a type checking bug where destructing can cause the type alias inliner to back off, which can cause further failures.
If there is a goog.require for an ES6 module that has a destructuring LHS, rewrite it early rather than letting the rewrite destructing pass handle it. The extra variable that pass creates causes the inliner to back off. That variable isn't needed as these getters are guaranteed side effect free.

Note that this isn't an issue for Closure modules now as they inline ALL require'd values, which is technically wrong. That ended up covering up this bug in the type checker.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=208551472