Use the module loader to resolve paths when using the regex based dependency parser. Be sure to pass the module loader when parsing.
This was an issue in closure_test_suite with CHECK_ONLY if the test used a []/ import path. It would not be able to resolve the import, and the imported file could be pruned as a result. It would only be pruned if it called declareNamespace - for some reason the compiler treats ES6 modules as entry points - but not if they declareNamespace. /shruggie

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=208686261