Better handling of anonymous classes in TypedScopeCreator
TypedScopeCreator was failing to infer the type of `this` and overriden method types in cases like

  use(class extends SomeSuperClass {
    f() {}
  });

because we didn't have a qualified name representing the method `f` in scope. Instead of requiring a qualified name, this CL just uses the property name `f` and the type on the class node.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=209077032