Backports an AngularJS 1.6 extern change to 1.5.
Changes the type of promise from an @constructor to an @interface. This CL doesn't change the runtime of AngularJS and doesn't matter much to the actual users of the externs. However this CL matters in turns of the .d.ts files generated by Clutz. With this change, the Clutz emitted .d.ts files won't have the "private noStructuralTypings_: any" brand in promise (because it's not a class anymore). Therefore it becomes assignable to a TS AngularJS promise (from angular.d.ts).

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=209843524