add JSType.restrictByPossiblyFalsy()
I intend to use this method in an upcoming change to add code
that creates new AST nodes from existing ones that already have type
information. It'll be used to calculate the right types for new
AND and OR nodes, for example.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=210971649