Fix GenerateExport's handling of class expressions
Previously when it saw a class expression that was not directly assigned to a qualified name, it would produce invalid code.

Now it does better in some cases for finding an appropiate qualified name for the class; if it can't find a qualified name it will report an error instead of silently producing bad code.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=211112904