Remove the Compiler `inputs` member variable.
The `inputs`, `modules` and `moduleGraph` compiler data structures overlap, which violates the SPOT rule and makes maintenance of the compiler's internal state more error-prone. This CL remedies part of the problem by removing `inputs` in favor of a `JSModuleGraph#getAllInputs` method, which returns an iterable over inputs in dependency order.

Note that it's already the case that `inputs` is in dependency order, since `initModules` and `rebuildInputsFromModules` (two of the three places where `inputs` is set, the third being `restoreState`) already initialize it in that manner. Thus, in most places where `inputs` occurs, we can simply replace it by a `moduleGraph.getAllInputs()` call. The exceptions are the methods that mutate the module graph: `hoistExterns` and `hoistNoCompileFiles`. In both cases, iteration must be over a copy of `getAllInputs()` to prevent concurrent modification.

Also note that, in `orderInputs`, passing `inputs` as an argument into `JSModuleGraph#manageDependencies` is redundant, since the graph itself contains that information.

Finally, a couple of places verify that the compiler has been initialized by a null check on `inputs`. These must be replaced by a null check on `moduleGraph`.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=211408645