Remove the Compiler `modules` member variable.
This is a followup to CL 211408645. After this change, the module graph will be the single point of truth for the module and input lists.

Since `modules` is initialized to the same value that gets passed into the `JSModuleGraph` constructor, we can replace any reference to the former by a call to `moduleGraph.getModules()` (which now returns an iterable, rather than a list).

Some care must be taken on `restoreState`, which needs to update the names of the deserialized modules to match the preexisting ones. To avoid concurrent modification, a copy of the previous module list must be made.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=211544774