Do not remove empty files in ConvertToTypedInterface pass.
Currently, passing an empty file into ConvertToTypedInterface within a test results in a "deleted scope was not reported" crash, because the deletion is reported to the deleted node itself, not the parent. It doesn't seem to be possible to report to the parent (the root) because it is neither a script nor a function. Therefore I'm removing the code that deletes empty files.

One test has to change because the synthetic "fill files" created by the compiler to populate otherwise empty chunks are now visible in the IJS output. We might want to figure out a way to get rid of them in the future.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=211686919