Abstract out the error printing strategy as a new interface
This makes it so that one can create custom error printing set-ups by passing
a strategy to the standard error manager, rather than always needing to create
a subclass.  As an example, converted the JSON stream report generator from
a subclass of the standard error manger to an instance of the new interface.

The end goal is to support multiple different ErrorReportGenerators to run
within a single compiler, like for example to print out normally formatted
messages to the console and at the same time also print JSON formatted messages
to an extra output file.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=211713942