Roll forward of "Extend the CheckRedundantNullabilityModifier linter pass to also check for missing nullability modifiers according to the style guide."
To prevent spurious warnings on template parameter types (which are allowed to be unqualified), do not warn on type names that appear in @template annotations in the same file. Because the check is purely syntactical (no type information), this incurs in some false positives and negatives. However, they are rare enough that it seems worthwhile to keep this check in the linter.

Also take the opportunity to add "void" to the list of primitive type names.

*** Reason for rollback ***

Added special-casing of template parameter types to prevent spurious warnings.

*** Original change description ***

Automated g4 rollback of changelist 212058481.

*** Reason for rollback ***

The new linter will incorrectly warn on an unqualified template parameter, e.g.

    /** @constructor @template T */ function Foo() {}
    /** @param {T} x */ Foo.prototype.bar = function(x) {};

will warn that the T in @param needs a ! or ?.

Rolling back to prevent a bad linter release.

*** Original change description ***

Extend the CheckRedundantNullabilityModifier linter pass to also check for missing nullability...

***

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=212655267