Check iterator immediately before access.
It is not immediately obvious that the Iterator queried in `constructAddOrStringNode()` always has a next element to fetch, since the `hasNext()` check occurs outside the querying method. The refactoring pulls this check closer to the query, in order to clarify that the access is always safe.

Closes https://github.com/google/closure-compiler/pull/3066

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=213511535