Correct property (dis)ambiguation for dynamically created classes.
This change is necessary for type checking and property (dis)ambiguation to
work correctly when class transpilation is moved after type checking.

The core of the change is to expand the logic in ObjectType.isAmbiguousObject()
to return true for cases where the implementation of the object's class or one
of its superclasses isn't fully known.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=215136357