Export Polymer element methods using local property @export annotations instead of goog.exportProperty.
Background:
This is a follow up to a change which had previously switched
to using goog.exportProperty instead of our own externs interface.
That change did fix two of our bugs for class-based code, but not legacy code,
and also caused a regression in user bundle size (see attached bugs).
The new approach in this change fixes the original bugs for both styles of code,
and also restores bundle size back to where it was before.

For more context, please see the JSDoc on the createExportsAndExterns function,
where I've documented all the different approaches I looked at,
and why this one seems to be the best.
(I'm also totally open to other approaches of course).

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=216623465