Fix flakiness of SymbolTableTest.
The flakiness caused by window.foo; being declared both in externs and main code and compiler doesn't know which one is declaration and which one is "usage" and depending on the order of processing they both might be used as declaration.

By changing test to use "window.foo = 1;" the compiler will treat it as usage. And that's how that variable is going to be used in actual code anyway.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=218450328