Automated g4 rollback
*** Reason for rollback ***

Breaks pantheon

*** Original change description ***

Fix aliasing of typedefs

Adds a new AST annotation to keep track of any typedef types defined by a definition site. This is different from the type of the value, in that e.g.

```
/** @typedef {string} */
ns.MyString;
```

calling `getJSType()` on the `ns.MyString` qualified name node returns the undefined type, since `ns.MyString` has a runtime value of undefined, but the type *name* represents a string, which we store in `Node#getTypedefTypeProp`. This is necessary so that typedefs can be dis...

***

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=219737644