Fix compiler crash during a special case with multi-line template literals.
The error occurs when printing the code using the CompactCodePrinter, when the multi-line template literal falls on the last line. In the CompactCodePrinter, there is logic to combine the last line with the line above, but this breaks if the "last line" contains line breaks (within template literals). The CodePrinter will mistakenly think that the last line is longer than it actually is.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=220862628