Expands `EqCache` to also cache comparison results of `TemplateTypeMap`s.
This enables accurate recursive comparisons of `TemplateTypeMap` by moving the "is recursive" state out of the map object. Storing that state in the map caused false positives when checking for recursion in some situations (when the check was triggered through an unexpected codepath). This solution is essentially the same one that was applied to fix this issue for `JSType`s.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=221302551