Treat destructured parameters as inferred if their type is unknown in TypedScopeCreator
This gets rid of some unknown typed parameters that were coming from looking up properties on unresolved NamedTypes. It's not ideal: we'd still prefer that these parameters have a known declared type, not known inferred type.

This is meant to be a temporary solution ; in the long term, we want to switch
to annotating types for individual destructuring parameters and ban giving a
single type for the entire parameter.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=221355865