Roll RemoveUnusedPolyfills into RemoveUnusedCode
By taking advantage of RemoveUnusedCode's better logic, we are able to do significantly better removal than we could before (in particular, removing the Set polyfill no longer leaves the Map and WeakMap polyfills behind).  This introduces a direct dependency on type information into RemoveUnusedCode where previously it relied entirely on disambiguation for type awareness.  Without type information, polyfills may still be removed, but the accuracy is poorer (unused polyfills may not be removed, and used polyfills that are hidden on aliases may be incorrectly removed).

This CL does not actually switch over to the new logic in DefaultPassConfig yet.  We will do that in a follow-up CL to minimize the size of potentially breaking changes in case of rollbacks.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=225476066