Automated g4 rollback of changelist 228183336.
*** Reason for rollback ***

Broke some internal runtime tests

*** Original change description ***

Back off less on conditional let/const declarations in LiveVariablesAnalysis

Given a conditionally executed var statement (e.g. for (var x of someArr) {}), it is possible that 'x' was previously assigned to, so 'x' is possibly live prior to the for loop. Normalize does not guarantee that a 'var' is never assigned to before the first 'var' statement.

However, given a conditional let/const declaration, it's safe to say that the variable was 'dead' prior to its declaration because it cannot be as...

***

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=228241572