Automated g4 rollback.
*** Reason for rollback ***

This breaks the Google Transparency Report.

The Google TR is currently crashing with the error "google is not defined" in compiled mode.

This problem will happen on any website that uses the standard [] externs rule, and has any dependency that does a goog.provide for google.*. This is because the common externs has an extern for
"var google = {}". In the case of Transparency Report, it's because we depend on the public Google Charts API, which provides the namespace 'google.charts'.

*** Original change description ***

Remove a little known feature from RescopeGlobalSymbols that rewrites most extern variables from e.g. `document` to `window.document`.

This was added (by me) to allow for chunks to be applied to multiple windows. Nobody is using it that way, and our current usage would like to avoid the overhead in bytes and runtime cost.

Supersedes []

***

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=228568960