Improve suspicious code checking of logical operators
Adds a check for the left-hand side of a logical operator having a known value.  It doesn't actually matter which value: if the left-hand side of an AND is always falsy then the right-hand side is dead, and if it's always truthy then its actual value is always discarded anyway.  The only semi-valid usage would be if there were a side effect, and in that case it's still clearer to move the expression to a previous statement, or (if more deeply nested) to use a comma operator.  The rationale for OR is the same, only reversed.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=228981771