Add $jscomp.inherits calls for interfaces.
We have to preserve the semantics of the uncompiled `extends`, or make this an error to compile. I'm fine with either.

How this was caught:

Someone is trying to mock an subinterface in their test that is written as a single inheritance chain of ES6 classes. This works untranspiled. Transpiled it won't pick up any methods of the superinterfaces because we don't add a $jscomp.inherits call.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=229405099