Set compiler on fill files for empty modules.
This is pre-work for an upcoming change, as without this tests fail as the compiler
is needed when getting dependency info.

The other option would be to explicitly skip fill files in hoistNoCompileFiles,
let me know if that is preferred.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=231500953