Fix infinite recursion in recursive-typedef subtype check
Recursive typedefs are not supposed to be possible, but there's a way to create them using forward-referenced named types. If such a recursive type is checked against a separate-but-equivalent @record or @typedef type, the type checker currently blows up with infinite recursion. This can be prevented by taking advantage of the ImplCache in PrototypeObjectType#isSubtype.

Doing this correctly requires carefully threading the needle through the various special cases and overrides of isSubtype, particularly due to a separate structural-subtype check that sometimes runs in addition to the version in PrototypeObjectType - if it reports (and caches) a different result then we end up with problems in union coalescing (either under- or over-collapsing), missing property detection, or various other places.
This CL attempts to consolidate the two structural checks into one by adding some extra dimensions to the check: (1) whether the supertype is an inline record literal or not (to determines which properties to match) and (2) whether to require explicitly voidable supertype props to be present in the subtype.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=232600301