Unconditionally extern all symbols used in runtime libraries
Removes the 'declare' pragma that was previously used in global.js, but was implicitly needed by other runtime libraries as well - though this fact was papered over by the Es6ExternsCheck.  Instead, we maintain a list of symbols in VarCheck that will be declared in externs at the end of the pass if they have not been encountered yet.  This is necessary because runtime libraries are injected *after* the first run of VarCheck, so any undeclared (normally externed) symbols in those files will cause undeclared variables that would crash later passes.

This approach is cleaner than conditioanlly declaring the symbols because (a) it avoids duplicate declarations, and (b) it leads to consistent behavior regardless of whether or not certain runtime libraries are included.

This change unblocks removing Es6ExternsCheck, ensuring that the compiler will not crash if no externs are provided.  Type checking runtime libraries without externs is not yet supported.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=232607582