All Refs created by their owning Name
This change enforces that GlobalNamespace.Ref objects will always be created by
the GlobalNamespace.Name object to which they belong.

This allows us to enforce several invariants such as:
* Ref.name will always == the owning Name object
* Ref.node will never be null when initially created except in unit tests
* Ref.scope will never be null except in unit tests

It also allows Name to control the creation of "twin" Refs, which
have several restrictions and are supposed to be the only way for two Refs
to refer to the same Node.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=232874810