Fix contentEditable and isContentEditable extern definitions.
contentEditable has always been a string and not a boolean, and is being extended with new values in a W3C proposal, which now fail the check. Making it a string breaks too much code, so we'll type it as (string|boolean) for the time being to aid the migration.

isContentEditable, on the other hand, is a boolean but was defined without a type.

Closes https://github.com/google/closure-compiler/pull/3228.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=234873466