Remove Es6ExternsCheck
It was only needed to prevent confusing compiler crashes in NTI.  Classic type inference builds the necessary information directly into the type registry so it will not crash.

This should have no visible effect on any existing working builds.  The worst reasonable unintended consequence is either (1) code that was failing with this error will now fail with a compiler crash, or (2) code that was failing with this error will now succeed to compile but produce invalid results.  I do not believe either of these to be remotely likely, but even if it did happen it would not be detectable with a presubmit test, since digging through the already-failing builds is not feasible.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=235815511