Make DiagnosticGroups.LINT_CHECKS disjoint from every other DiagnosticGroup.
This is necessary to avoid surprising results when a DiagnosticGroup that intersects with LINT_CHECKS is enabled, as it will cause all of the other LINT_CHECKS to run too. This is especially confusing in conjunction with --strict, as it causes linter changes to introduce regressions due to the presence of a seemingly unrelated flag.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=236923617