Update `GlobalNamespace` and `CollapseProperties` to consider object rest and spread.
The added support is minimal. In some cases we could probably do better, but this feature isn't expected to come up often so isn't currently worth the complexity.

NEW: Includes a fix for crashes when the LHS of an assignment is a cast.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=236934654