Invert the sense of the bits inside `SideEffectFlags` so that they're easier to understand.
The original implementation essentially used a "0 means on" encoding. This was because the default value of an int `Node` prop is 0, and the safe assumption for side-effects is that they're all on. This change switches things to invert the bits during set/get so that the default value (0) still means "all side-effects", but all the rest of the logic is simpler.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=237853763