Migrate the standardized Console to a separate top-level extern.
TThe purpose is to allow projects that only want to support the
standardized part of the API to pull in that single extern.
Currently google/elemental2 has copy pasted the relevant section
into the downstream project. Non-standardized aspects will
continue to be present in webkit_dom.js. This patch also renames
some parameter names to align with the names in the spec and
orders the declarations as they appear in the spec.