Modifies property override validation to cover purely prototypal inheritance.
This is required for validation of class-side overrides, to ensure that static members of an ES6 syntax subclass are compatible with any superclass definitions. This applied to ES6 syntax interfaces too, as well as any type with a known prototypal heritage.

The errors generated by this check are included in a new diagnostic group 'checkPrototypalTypes'.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=240827487