Explicitly mark const values
If a value is defined as "const" in the WebIDL or the spec it
should be marked as such. The advantage is that this flows into
downstream consumers such as google/elemental2 allowing those
projects to define values as constants.

Closes https://github.com/google/closure-compiler/pull/3276.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=240843054