Enforce that `PureFunctionIdentifier` runs over only normalized ASTs.
This is a step towards running `PureFunctionIdentifer` using `OptimizeCalls` which also requires normalization to have run.

Enforcing normalization requires updating tests for passes that depend on pure function identification to also perform normalization. An upshot of this is that some test cases become redundant (and are deleted) when this is enforced.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=241393703